+++ orchestrator/schemas/__init__.py
@@ -11,7 +11,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from orchestrator.schemas.engine_settings import EngineSettingsBaseSchema, EngineSettingsSchema, GlobalStatusEnum
+from orchestrator.schemas.engine_settings import (
+    EngineSettingsBaseSchema,
+    EngineSettingsSchema,
+    GlobalStatusEnum,
+    WorkerStatus,
+)
 from orchestrator.schemas.fixed_input import FixedInputConfigurationSchema, FixedInputSchema
 from orchestrator.schemas.problem_detail import ProblemDetailSchema
 from orchestrator.schemas.process import (
@@ -60,6 +65,7 @@ __all__ = (
     "SubscriptionIdSchema",
     "ResourceTypeSchema",
     "ResourceTypeBaseSchema",
+    "WorkerStatus",
     "WorkflowSchema",
     "WorkflowWithProductTagsSchema",
 )