+++ orchestrator/schemas/engine_settings.py
@@ -27,9 +27,17 @@ class EngineSettingsBaseSchema(OrchestratorBaseModel):
     global_lock: bool
 
 
+class WorkerStatus(OrchestratorBaseModel):
+    executor_type: str
+    number_of_workers_online: int = 0
+    number_of_queued_jobs: int = 0
+    number_of_running_jobs: int = 0
+
+
 class EngineSettingsSchema(EngineSettingsBaseSchema):
     global_status: Optional[GlobalStatusEnum]
     running_processes: int
+    worker_status: Optional[WorkerStatus]
 
     class Config:
         orm_mode = True